Extract ModelStub into a public class
