+++ src/test/java/seedu/address/model/ModelStub.java
@@ -1,0 +1,207 @@
+package seedu.address.model;
+
+import java.nio.file.Path;
+import java.util.Comparator;
+import java.util.function.Predicate;
+
+import javafx.collections.ObservableList;
+import seedu.address.commons.core.GuiSettings;
+import seedu.address.model.meeting.Meeting;
+import seedu.address.model.person.Person;
+import seedu.address.model.reminder.Reminder;
+import seedu.address.model.sale.Sale;
+import seedu.address.model.tag.Tag;
+
+/**
+ * A default model stub that have all of the methods failing.
+ */
+public class ModelStub implements Model {
+    @Override
+    public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ReadOnlyUserPrefs getUserPrefs() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public GuiSettings getGuiSettings() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setGuiSettings(GuiSettings guiSettings) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public Path getAddressBookFilePath() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setAddressBookFilePath(Path addressBookFilePath) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addPerson(Person person) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setAddressBook(ReadOnlyAddressBook newData) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ReadOnlyAddressBook getAddressBook() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasPerson(Person person) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasContactTag(Tag tag) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasSaleTag(Tag tag) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addContactTag(Tag tag) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void editContactTag(Tag target, Tag editedTag) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void editSaleTag(Tag target, Tag editedTag) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deleteContactTag(Tag target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deleteSaleTag(Tag target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public int findByContactTag(Tag target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public String findBySaleTag(Tag target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public String listTags() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Tag> getContactTagList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Tag> getSaleTagList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deletePerson(Person target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setPerson(Person target, Person editedPerson) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Person> getFilteredPersonList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void updateFilteredPersonList(Predicate<Person> predicate) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Person> getSortedPersonList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    public void updateSortedPersonList(Comparator<Person> comparator) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Meeting> getSortedMeetingList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasMeeting(Meeting meeting) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deleteMeeting(Meeting target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addMeeting(Meeting meeting) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Reminder> getSortedReminderList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasReminder(Reminder reminder) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deleteReminder(Reminder target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addReminder(Reminder reminder) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addSaleToPerson(Person person, Sale sale) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void removeSaleFromPerson(Person person, Sale sale) {
+        throw new AssertionError("This method should not be called.");
+    }
+}
